Animator + AnimatorHandler Implementation + TextureHandler Function Changes
Big Changes - Reduced parameter amounts for TextureHandler methods (no longer passing in renderer or camera), this means the TextureHandler is yet again in-complete.

Thinking of doing some sort of SceneHandler which holds a collection of Scene objects that define the overall "game",  the scene manager will have a ref to currently active camera that we can use in texture handler methods and engine class will hold a ref to renderer which we will also use in texture handler methods.